feature(bit): add copied header for <bit>
This header is already available as part of a freestanding
implementation of the standard library; however, it includes some
alternative portions for a freestanding implementation that conflict
with other headers copied into extended-freestanding. "bit" is added to
extended-freestanding to make tweaks to make it work with the other
headers of extended-freestanding.